Reorganized PCSE code
- Moved all WOFOST code to a separate pcse.crop.wofost subpackage
- Deleted some files that were not used anymore:
  * taskmanager.py
  * rerunner.py
  * everything related to npk dynamics
