+++ pcse/crop/__init__.py
@@ -2,15 +2,8 @@
 # Copyright (c) 2004-2024 Wageningen Environmental Research, Wageningen-UR
 # Allard de Wit (allard.dewit@wur.nl), March 2024
 from . import abioticdamage
-from . import assimilation
-from . import evapotranspiration
-from . import partitioning
-from . import phenology
-from . import respiration
-from . import root_dynamics
-from . import stem_dynamics
-from . import storage_organ_dynamics
-from . import leaf_dynamics
+from .wofost import assimilation, leaf_dynamics, partitioning, phenology, respiration, root_dynamics, stem_dynamics, \
+    storage_organ_dynamics, evapotranspiration
 from . import lintul3
 from . import lingra
 from . import lingraN
