--- pcse/crop/nutrients/__init__.py
@@ -1,2 +1,0 @@
-from .n_stress import N_Stress
-from .n_demand_uptake import N_Demand_Uptake
