+++ pcse/tests/test_evapotranspiration.py
@@ -4,7 +4,7 @@
 import unittest
 from datetime import date
 
-from ..crop.evapotranspiration import Evapotranspiration
+from pcse.crop.wofost.evapotranspiration import Evapotranspiration
 from ..base import VariableKiosk, ParameterProvider
 from .test_data import pot_evtra_testdata, wl_evtra_testdata1,\
     wl_parvalue_dict1, wl_evtra_testdata2, wl_parvalue_dict2
